nlp-service/pyproject.toml
[tool.poetry]
name = "nlpservice"
version = "0.1.0"
description = "A Python service that wraps various ML libraries"
authors = ["Aaron <AaronBatilo@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.28"
protobuf = "^3.11"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.28"

[tool.poetry.scripts]
entrypoint = 'nlpservice:server.serve'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
